+++ objc2/CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
       msg_send_id![NSObject::alloc(), init]
   };
   ```
+* Add `Class::class_method`.
 
 ### Changed
 * Allow other types than `&Class` as the receiver in `msg_send_id!` methods